# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API untuk manajemen turnamen esports Mobile Legends",
        "title": "EMBECK API - Turnamen Esports Management",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/rrq-dev/EMBECK",
            "email": "support@embeck.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/admin/matches": {
            "get": {
                "description": "Mendapatkan daftar semua pertandingan dengan filter tournament_id opsional",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matches"
                ],
                "summary": "Get All Matches",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tournament ID filter",
                        "name": "tournament_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Match"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Membuat pertandingan baru untuk turnamen",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matches"
                ],
                "summary": "Create New Match",
                "parameters": [
                    {
                        "description": "Match data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MatchRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.MatchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/admin/matches/{id}": {
            "get": {
                "description": "Mendapatkan detail pertandingan berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matches"
                ],
                "summary": "Get Match By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Match ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Match"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Memperbarui detail pertandingan termasuk input skor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matches"
                ],
                "summary": "Update Match",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Match ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Match data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.MatchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Menghapus pertandingan dari database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Matches"
                ],
                "summary": "Delete Match",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Match ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.MatchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Match": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "match_date": {
                    "type": "string"
                },
                "match_time": {
                    "type": "string"
                },
                "result_team_a_score": {
                    "type": "integer"
                },
                "result_team_b_score": {
                    "type": "integer"
                },
                "round": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "team_a_id": {
                    "type": "string"
                },
                "team_b_id": {
                    "type": "string"
                },
                "tournament_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "winner_team_id": {
                    "type": "string"
                }
            }
        },
...
        "model.MatchRequest": {
            "type": "object",
            "required": [
                "match_date",
                "match_time",
                "round",
                "status",
                "team_a_id",
                "team_b_id",
                "tournament_id"
            ],
            "properties": {
                "location": {
                    "type": "string"
                },
                "match_date": {
                    "type": "string"
                },
                "match_time": {
                    "type": "string"
                },
                "result_team_a_score": {
                    "type": "integer"
                },
                "result_team_b_score": {
                    "type": "integer"
                },
                "round": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "scheduled",
                        "ongoing",
                        "completed",
                        "cancelled"
                    ]
                },
                "team_a_id": {
                    "type": "string"
                },
                "team_b_id": {
                    "type": "string"
                },
                "tournament_id": {
                    "type": "string"
                },
                "winner_team_id": {
                    "type": "string"
                }
            }
        },
        "model.MatchResponse": {
            "type": "object",
            "properties": {
                "match_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}